pull some utility functions out of connectAdvanced into their own utils files
