+++ src/index.js
@@ -1,5 +1,13 @@
 import Provider from './components/Provider'
 import connect from './components/connect'
-import connectAdvanced, { dispatchable } from './components/connectAdvanced'
+import connectAdvanced from './components/connectAdvanced'
+import createShallowEqualSelector from './utils/createShallowEqualSelector'
+import dispatchable from './utils/dispatchable'
 
-export { Provider, connect, connectAdvanced, dispatchable }
+export {
+  Provider,
+  connect,
+  connectAdvanced,
+  createShallowEqualSelector,
+  dispatchable 
+}
