+++ src/utils/createShallowEqualSelector.js
@@ -1,0 +1,4 @@
+import { createSelectorCreator, defaultMemoize } from 'reselect'
+import shallowEqual from './shallowEqual'
+
+export default createSelectorCreator(defaultMemoize, shallowEqual)
